chore: remove GoCardless integration
It doesn't support viewing transactions created after linking the bank
account, meaning it is only useful when importing data for the first
time. I plan on adding CSV import later on, so this scenario would be
covered by it.

More advantages:
- Less code to maintain
- Removes a middleman that can preview bank info

Sad to see all the effort of bank selection card colors go to waste.