feat: show overlay and labels when clicking on home page FAB
This also fixes the varying padding for mobile and desktop platforms. By
using `MaterialTapTargetSize.padded`, the resulting FAB has the same
width on all platforms, which makes calculating offsets easier.

The `flutter_expandable_fab` package contains a bug that I fixed in my
fork and created an issue for it here
https://github.com/zuvola/flutter_expandable_fab/issues/30